chore(auth): rename category level result.ts to output.ts
