+++ packages/auth/src/types/outputs.ts
